xds: migrate to restructured xds loadbalancer
