+++ xds/src/main/java/io/grpc/xds/XdsComms2.java
@@ -36,7 +36,7 @@ import io.grpc.Status;
 import io.grpc.SynchronizationContext.ScheduledHandle;
 import io.grpc.internal.BackoffPolicy;
 import io.grpc.stub.StreamObserver;
-import io.grpc.xds.LookasideChannelLb.AdsStreamCallback2;
+import io.grpc.xds.LookasideChannelLb.AdsStreamCallback;
 import java.util.concurrent.TimeUnit;
 import javax.annotation.CheckForNull;
 
@@ -63,7 +63,7 @@ final class XdsComms2 {
     static final String EDS_TYPE_URL =
         "type.googleapis.com/envoy.api.v2.ClusterLoadAssignment";
 
-    final AdsStreamCallback2 adsStreamCallback2;
+    final AdsStreamCallback adsStreamCallback;
     final StreamObserver<DiscoveryRequest> xdsRequestWriter;
     final Stopwatch retryStopwatch = stopwatchSupplier.get().start();
 
@@ -90,7 +90,7 @@ final class XdsComms2 {
                         value.getResources(0).unpack(ClusterLoadAssignment.class);
                   } catch (InvalidProtocolBufferException | RuntimeException e) {
                     cancelRpc("Received invalid EDS response", e);
-                    adsStreamCallback2.onError();
+                    adsStreamCallback.onError();
                     scheduleRetry();
                     return;
                   }
@@ -98,7 +98,7 @@ final class XdsComms2 {
                   helper.getChannelLogger().log(
                       ChannelLogLevel.DEBUG,
                       "Received an EDS response: {0}", clusterLoadAssignment);
-                  adsStreamCallback2.onEdsResponse(clusterLoadAssignment);
+                  adsStreamCallback.onEdsResponse(clusterLoadAssignment);
                 }
               }
             }
@@ -116,7 +116,7 @@ final class XdsComms2 {
                     if (cancelled) {
                       return;
                     }
-                    adsStreamCallback2.onError();
+                    adsStreamCallback.onError();
                     scheduleRetry();
                   }
                 });
@@ -168,8 +168,8 @@ final class XdsComms2 {
     boolean cancelled;
     boolean closed;
 
-    AdsStream(AdsStreamCallback2 adsStreamCallback2) {
-      this.adsStreamCallback2 = adsStreamCallback2;
+    AdsStream(AdsStreamCallback adsStreamCallback) {
+      this.adsStreamCallback = adsStreamCallback;
       this.xdsRequestWriter = AggregatedDiscoveryServiceGrpc.newStub(channel).withWaitForReady()
           .streamAggregatedResources(xdsResponseReader);
 
@@ -190,7 +190,7 @@ final class XdsComms2 {
     }
 
     AdsStream(AdsStream adsStream) {
-      this(adsStream.adsStreamCallback2);
+      this(adsStream.adsStreamCallback);
     }
 
     // run in SynchronizationContext
@@ -208,13 +208,13 @@ final class XdsComms2 {
    * Starts a new ADS streaming RPC.
    */
   XdsComms2(
-      ManagedChannel channel, Helper helper, AdsStreamCallback2 adsStreamCallback2,
+      ManagedChannel channel, Helper helper, AdsStreamCallback adsStreamCallback,
       BackoffPolicy.Provider backoffPolicyProvider, Supplier<Stopwatch> stopwatchSupplier) {
     this.channel = checkNotNull(channel, "channel");
     this.helper = checkNotNull(helper, "helper");
     this.stopwatchSupplier = checkNotNull(stopwatchSupplier, "stopwatchSupplier");
     this.adsStream = new AdsStream(
-        checkNotNull(adsStreamCallback2, "adsStreamCallback2"));
+        checkNotNull(adsStreamCallback, "adsStreamCallback"));
     this.backoffPolicyProvider = checkNotNull(backoffPolicyProvider, "backoffPolicyProvider");
     this.adsRpcRetryPolicy = backoffPolicyProvider.get();
   }
