+++ xds/src/test/java/io/grpc/xds/XdsLocalityTest.java
@@ -20,12 +20,13 @@ import static com.google.common.truth.Truth.assertThat;
 
 import com.google.common.testing.EqualsTester;
 import io.envoyproxy.envoy.api.v2.core.Locality;
+import io.grpc.xds.ClusterLoadAssignmentData.XdsLocality;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 
 /**
- * Unit tests for {@link XdsLocality}.
+ * Unit tests for {@link ClusterLoadAssignmentData.XdsLocality}.
  */
 @RunWith(JUnit4.class)
 public class XdsLocalityTest {
